requirements.txt
streamlit>=1.35.0
langchain>=0.1.17
langchain-community>=0.0.21
chromadb>=0.4.24
sentence-transformers>=2.2.2
transformers>=4.41.0
torch>=2.0.0
tqdm>=4.65.0
accelerate>=0.30.0
python-dotenv>=1.0.1
pypdf>=3.17.1
